+++ src/app/(site)/_components/home-header.tsx
@@ -1,0 +1,16 @@
+import { Header } from '~/components/header';
+import { ListItem } from './list-item';
+
+export function HomeHeader() {
+  return (
+    <Header>
+      <div className="mb-2">
+        <h1 className="text-3xl font-semibold text-white">
+          Fountain House Music
+        </h1>
+
+        <ListItem image="/images/liked.jpeg" name="Liked Songs" href="liked" />
+      </div>
+    </Header>
+  );
+}
