+++ src/app/(site)/page.tsx
@@ -1,6 +1,5 @@
-import { Header } from '~/components/header';
 import { getAlbums } from '~/server/actions/album/get-albums';
-import { ListItem } from './_components/list-item';
+import { HomeHeader } from './_components/home-header';
 import { PageContent } from './_components/page-content';
 
 export const revalidate = 0;
@@ -10,19 +9,7 @@ export default async function Home() {
 
   return (
     <div className="w-full h-full overflow-hidden overflow-y-auto rounded-lg bg-neutral-900">
-      <Header>
-        <div className="mb-2">
-          <h1 className="text-3xl font-semibold text-white">
-            Fountain House Music
-          </h1>
-
-          <ListItem
-            image="/images/liked.jpeg"
-            name="Liked Songs"
-            href="liked"
-          />
-        </div>
-      </Header>
+      <HomeHeader />
       <div className="px-6 mt-2 mb-7">
         <div className="flex items-center justify-between">
           <h1 className="text-2xl font-semibold text-white">Newest Albums!</h1>
